Исправлены замечания: + @Controller висит в воздухе + В слое репозитория ставится другая аннотация над классом. + @Transactional(readOnly = true) над классом убрать и перенести транзакции в сервис. + Удали весь закомментированный код и комментарии. + В сущностях добавить иквалс и хешкод. + Добавить слой сервиса для RoleRepository, UserRepository и уже в класс контроллера инжекть через интерфейс сервиса. На прямую репозиторий в слой контроллера не инжектят! + Используй или дао, или репозиторий. + private UserDbDAO userDbDAO; дао в слой контроллера инжектится через сервис + Зачем тебе 2 сущности User, сделай одну. +Для UserService добавить интерфейс и его инжектить в контроллер.

+++ src/main/java/mvc/spring/security/repositories/RoleRepository.java
@@ -5,8 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
-
-    // использую этот метод для реализации автозаполнения таблицы ролей в классе RoleInitializer:
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Role findByName(String name);
 }
